update to follow recommendations from PR comments
