+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/localization/Localizer.java
@@ -7,16 +7,19 @@ import com.acmerobotics.roadrunner.PoseVelocity2d;
  * Interface for localization methods.
  */
 public interface Localizer {
-    /**
-     * Updates the position estimate.
-     * @param pose the current position estimate
-     * @return the updated position estimate
-     */
-    Pose2d updatePositionEstimate(Pose2d pose);
+    void setPose(Pose2d pose);
 
     /**
-     * Updates the velocity estimate.
-     * @return the updated velocity estimate
+     * Returns the current pose estimate.
+     * NOTE: Does not update the pose estimate;
+     * you must call update() to update the pose estimate.
+     * @return the Localizer's current pose
      */
-    PoseVelocity2d updateVelocityEstimate();
+    Pose2d getPose();
+
+    /**
+     * Updates the Localizer's position estimate.
+     * @return the Localizer's current velocity estimate
+     */
+    PoseVelocity2d update();
 }
