+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/localization/TwoDeadWheelLocalizer.java
@@ -43,8 +43,9 @@ public final class TwoDeadWheelLocalizer implements Localizer {
 
     private double lastRawHeadingVel, headingVelOffset;
     private boolean initialized;
+    private Pose2d pose;
 
-    public TwoDeadWheelLocalizer(HardwareMap hardwareMap, IMU imu, double inPerTick) {
+    public TwoDeadWheelLocalizer(HardwareMap hardwareMap, IMU imu, double inPerTick, Pose2d pose) {
         // TODO: make sure your config has **motors** with these names (or change them)
         //   the encoders should be plugged into the slot matching the named motor
         //   see https://ftc-docs.firstinspires.org/en/latest/hardware_and_software_configuration/configuring/index.html
@@ -59,16 +60,25 @@ public final class TwoDeadWheelLocalizer implements Localizer {
         this.inPerTick = inPerTick;
 
         FlightRecorder.write("TWO_DEAD_WHEEL_PARAMS", PARAMS);
+
+        this.pose = pose;
     }
 
     @Override
-    public Pose2d updatePositionEstimate(Pose2d pose) {
-        return pose.plus(calculatePoseDelta().value());
+    public void setPose(Pose2d pose) {
+        this.pose = pose;
     }
 
     @Override
-    public PoseVelocity2d updateVelocityEstimate() {
-        return calculatePoseDelta().velocity().value();
+    public Pose2d getPose() {
+        return pose;
+    }
+
+    @Override
+    public PoseVelocity2d update() {
+        Twist2dDual<Time> delta = calculatePoseDelta();
+        pose = pose.plus(delta.value());
+        return delta.velocity().value();
     }
 
     public Twist2dDual<Time> calculatePoseDelta() {
